actually die on worker panics
